third_party/blink/web_tests/platform/linux-chrome/external/wpt/wai-aria/role/list-roles-expected.txt

This is a testharness.js-based test.
[FAIL] orphan p with listitem role
  assert_equals: <p role="listitem" data-testname="orphan p with listitem role" data-expectedrole="paragraph" class="ex">x</p> expected "paragraph" but got "listitem"
Harness: the test ran to completion.
